pyproject.toml
[project]
name = "crypto-magneto-prod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.8",
    "django-compressor>=4.5.1",
    "django-imagekit>=6.0.0",
    "django-redis>=6.0.0",
    "httpx>=0.28.1",
    "logfire[django]>=4.14.2",
    "psycopg[binary]>=3.2.12",
    "pycountry>=24.6.1",
    "pydantic>=2.12.4",
    "rcssmin>=1.1.2",
    "redis>=7.0.1",
    "rjsmin>=1.2.2",
]
